wip; impl chdir; first stab at projection
