+++ lib/index.js
@@ -60,7 +60,9 @@ function FSTree(options) {
     validateSortedUnique(entries);
   }
 
-  this.entries = entries;
+  this._entries = entries;
+  this._parent = options.parent || null;
+
   if ('root' in options) {
     let root = options.root;
     if (typeof root !== 'string' || !path.isAbsolute(root)) {
@@ -68,11 +70,39 @@ function FSTree(options) {
     }
 
     this.root = path.normalize(options.root + path.sep);
+    this.cwd = options.cwd || ''; // relative to this.root
   }
 
-  this.start();
-  this._changes = [];
+  // TODO: share state?
+  // if not we probably don't want `chdir` to clear changes (as it would do
+  // indirectly via start)
+  if (!this._parent) {
+    this.__changes = [];
+    this.start();
+  }
 }
+
+FSTree.prototype = {
+  get entries() {
+    return this._parent ? this._parent.entries : this._entries;
+  },
+
+  get _changes() {
+    return this._parent ? this._parent._changes : this.__changes;
+  },
+
+  get _relativePathToChange() {
+    return this._parent ? this._parent._relativePathToChange : this.__relativePathToChange;
+  },
+
+  set _relativePathToChange(value) {
+    if (this._parent) {
+      this._parent._relativePathToChange = value;
+    } else {
+      this.__relativePathToChange = value;
+    }
+  }
+};
 
 FSTree.fromPaths = function(paths, options) {
   if (typeof options !== 'object') { options = {}; }
@@ -126,7 +156,7 @@ FSTree.prototype.addPaths = function(paths, options) {
 };
 
 FSTree.prototype.start = function() {
-  this._changes = [];
+  this._changes.splice(0, this._changes.length);
   this._relativePathToChange = Object.create(null);
   this._state = STARTED;
 };
@@ -135,8 +165,12 @@ FSTree.prototype.stop = function() {
   this._state = STOPPED;
 };
 
+FSTree.prototype._normalizePath = function(relativePath) {
+  return chompPathSep(`${this.cwd}${relativePath}`);
+};
+
 FSTree.prototype.findByRelativePath = function(relativePath) {
-  relativePath = chompPathSep(relativePath);
+  relativePath = this._normalizePath(relativePath);
 
   // TODO: experiment with binary search since entries are sorted
   for (var i = 0; i < this.entries.length; i++){
@@ -164,7 +198,7 @@ FSTree.prototype.match = function(globs) {
 };
 
 FSTree.prototype.changes = function(globs) {
-  var changes = this._changes || this._initialChanges; // TODO: order correctly
+  var changes = this._changes;
   if (arguments.length > 0) {
     var included = new MatcherCollection(globs.include);
     var excluded;
@@ -181,13 +215,26 @@ FSTree.prototype.changes = function(globs) {
   }
 };
 
+FSTree.prototype.chdir = function(relativePath, options) {
+  let cwd = relativePath === '' ? '' : `${chompPathSep(relativePath)}/`;
+
+  if (cwd === this.cwd) { return this; }
+
+  return new FSTree({
+    root: this.root,
+    cwd: cwd,
+    parent: this,
+  });
+};
+
 FSTree.prototype.readFileSync = function(relativePath, encoding) {
-  var entry = this.findByRelativePath(relativePath);
+  var result = this.findByRelativePath(relativePath);
+  var entry = result.entry;
 
   // if instead of this.root we asked the entry, we could emulate symlinks on
   // readFileSync. (there will be other things to do as well, for example
   // rmdir/unlink etc..
-  return fs.readFileSync(this.root + '/' + relativePath, encoding);
+  return fs.readFileSync(this.root + '/' + entry.relativePath, encoding);
 };
 
 FSTree.prototype._throwIfStopped = function(operation) {
@@ -202,7 +249,7 @@ FSTree.prototype.unlinkSync = function(relativePath) {
   var result = this.findByRelativePath(relativePath);
   var entry = result.entry;
 
-  fs.unlinkSync(this.root + '/' + relativePath);
+  fs.unlinkSync(this.root + '/' + entry.relativePath);
   this._track('unlink', entry);
   this._removeAt(result);
 };
@@ -213,7 +260,7 @@ FSTree.prototype.rmdirSync = function(relativePath) {
   var result = this.findByRelativePath(relativePath);
   var entry = result.entry;
 
-  fs.rmdirSync(this.root + '/' + relativePath);
+  fs.rmdirSync(this.root + '/' + entry.relativePath);
   this._track('rmdir', entry);
   this._removeAt(result);
 };
@@ -229,8 +276,9 @@ FSTree.prototype.mkdirSync = function(relativePath) {
     return;
   }
 
-  fs.mkdirSync(this.root + relativePath);
-  entry = new Entry(relativePath, 0, Date.now(), Entry.DIRECTORY_MODE, null);
+  let normalizedPath = this._normalizePath(relativePath);
+  fs.mkdirSync(`${this.root}${normalizedPath}`);
+  entry = new Entry(normalizedPath, 0, Date.now(), Entry.DIRECTORY_MODE, null);
 
   this._track('mkdir', entry);
   this._insertAt(result, entry);
@@ -262,8 +310,10 @@ FSTree.prototype.writeFileSync = function(relativePath, content, options) {
     };
   }
 
-  fs.writeFileSync(this.root + '/' + relativePath, content, options);
-  var entry = new Entry(relativePath, content.length, Date.now(), mode || 0, checksum);
+  let normalizedRelativePath = this._normalizePath(relativePath);
+
+  fs.writeFileSync(this.root + normalizedRelativePath, content, options);
+  var entry = new Entry(normalizedRelativePath, content.length, Date.now(), mode || 0, checksum);
   var operation = result.entry ? 'change' : 'create';
 
   this._track(operation, entry);
@@ -280,7 +330,8 @@ FSTree.prototype.symlinkSync = function(target, relativePath /*, type */) {
     return;
   }
 
-  symlinkOrCopy.sync(target, `${this.root}${relativePath}`);
+  let normalizedPath = this._normalizePath(relativePath);
+  symlinkOrCopy.sync(target, `${this.root}${normalizedPath}`);
 
   // TODO: do we need this pattern?  used in funnel
   //
@@ -303,7 +354,7 @@ FSTree.prototype.symlinkSync = function(target, relativePath /*, type */) {
 
   // TODO: should we read the file here so our entry has size, mode & checksum?
   // this turns 1 io -> 2 io (symlink -> symlink + stat)
-  let entry = new Entry(relativePath, 0, Date.now(), 0);
+  let entry = new Entry(normalizedPath, 0, Date.now(), 0);
   let operation = result.entry ? 'change' : 'create';
   this._track(operation, entry);
   this._insertAt(result, entry);
@@ -374,7 +425,9 @@ FSTree.prototype._removeAt = function(result) {
 };
 
 FSTree.prototype.filter = function(fn, context) {
-  return this.entries.filter(fn, context);
+  return this.entries.filter(e => {
+    return entryRelativePath(e).indexOf(this.cwd) === 0;
+  }).filter(fn, context);
 };
 
 FSTree.prototype.forEach = function(fn, context) {
