+++ package.json
@@ -7,8 +7,8 @@
     "lib"
   ],
   "scripts": {
-    "test": "mocha tests/",
-    "test:debug": "mocha debug tests/"
+    "test": "mocha --recursive tests/",
+    "test:debug": "mocha --recursive debug tests/"
   },
   "keywords": [
     "broccoli"
