+++ tests/fs-merge-tree-test.js
@@ -1,0 +1,7 @@
+const expect = require('chai').expect;
+
+describe('FSMergeTree', function() {
+  it('has tests', function() {
+    expect('assertions exist').to.eql(true);
+  });
+});
